// src/services/generated/models/flagType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * SEMD API
 * OpenAPI spec version: 1.0.0
 */

export type FlagType = typeof FlagType[keyof typeof FlagType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FlagType = {
  MALICIOUS: 'MALICIOUS',
  BENIGN: 'BENIGN',
} as const;