// src/services/generated/models/reportStatusType.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * SEMD API
 * OpenAPI spec version: 1.0.0
 */

export type ReportStatusType = typeof ReportStatusType[keyof typeof ReportStatusType];


export const ReportStatusType = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
} as const;